+++ Lab/code/clearNotes.js
