--- LabLoop/code/getDensity.js
@@ -1,2 +1,0 @@
-var clipSlot = new LiveAPI("live_set view highlighted_clip_slot");
-// Placeholder
