--- LabLoop/code/getFilePath.js
@@ -1,6 +1,0 @@
-outlets = 1;
-
-function bang() {
-    var patcherPath = this.patcher.filepath;
-    outlet(0, patcherPath);
-}
