--- LabLoop/node_content/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2021",
-    "module": "commonjs",
-    "declaration": true,
-    "noImplicitAny": false,
-    "strict": true,
-		"incremental": true,
-		"esModuleInterop": true,
-    "outDir": "dist"
-  },
-	"include": [
-		"./src"
-	]
-}
